+++ vpc-20160428/ChangeLog.md
@@ -1,3 +1,11 @@
+2026-09-15 Version: 7.3.1
+- Update API CreateVSwitch: add request parameters CidrMask.
+- Update API CreateVSwitch: add request parameters Ipv6CidrMask.
+- Update API DescribeRouteEntryList: add response parameters Body.TotalCount.
+- Update API GetVpcPrefixListAssociations: add response parameters Body.PrefixListAssociation.$.SuppressedList.
+- Update API ModifyVSwitchAttribute: add request parameters Ipv6CidrMask.
+
+
 2026-09-01 Version: 7.3.0
 - Support API ModifyVpcCidrBlock.